% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/625package.R
\name{survival.probs}
\alias{survival.probs}
\title{Ariana}
\usage{
square()
}
\arguments{
\item{x}{input value}
}
\value{
the square of x
}
\description{
Gets the square of a number
}
\examples{
square(3)

}
