% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/625package.R, R/survivalprobs.R
\name{survival.probs}
\alias{survival.probs}
\title{survival.probs}
\usage{
survival.probs(time, event)

survival.probs(time, event)
}
\arguments{
\item{x}{vector of times and vector of events}
}
\value{
vector of survival probabilities

vector of survival probabilities
}
\description{
Gets a vector of survival probabilities for each time point

Gets a vector of survival probabilities for each time point
}
\examples{
survival.probs()

time = c(3,2,5,4.1)
event = c(1,0,0,1,1)
survival.probs(time, event)

}
