% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalprobs.R
\name{survival.probs}
\alias{survival.probs}
\title{survival.probs}
\usage{
survival.probs(time, event)
}
\arguments{
\item{time}{vector of times}

\item{event}{vector of events}
}
\value{
vector of survival probabilities
}
\description{
Gets a vector of survival probabilities for each time point
}
\examples{
time = c(3,2,5,4,1)
event = c(1,0,0,1,1)
survival.probs(time, event)

}
